feat: 处理配置

+++ dev.sh
@@ -3,14 +3,28 @@
 export NODE_ENV=development
 export DATA_PATH=../data
 
-# 为整个文件夹添加操作权限
+# 为 nginx 目录设置适当的权限
 chmod -R 777 .
 
 # 创建必要的目录
 mkdir -p "${DATA_PATH}/nginx/conf"
 
-# 复制 nginx.conf 到目标目录
-cp ./config/nginx.conf "${DATA_PATH}/nginx/conf"
+# 检查 nginx.conf 文件是否存在
+if [ -f ./config/nginx.conf ]; then
+  # 复制文件到目标位置
+  cp ./config/nginx.conf "${DATA_PATH}/nginx/conf"
+
+  # 输出文件内容（可调试用）
+  echo "复制前的文件内容："
+  cat ./config/nginx.conf
+
+  echo "复制后的文件内容："
+  cat "${DATA_PATH}/nginx/conf/nginx.conf"
+
+else
+  echo "nginx.conf 文件不存在，无法复制。"
+  exit 1
+fi
 
 # 构建并启动开发环境
 docker compose -f docker-compose-dev.yml up -d
